Fixed a reliability problem with TimeSync events for end of bar during real time for time based bars. Now, if creating a time fails because the time isn't in the future, the Data object will update the bar's next start time and retry till it definitely creates a timer.

+++ Platform/TickZoomAPI1.0/Classes/Elapsed.cs
@@ -83,8 +83,8 @@ namespace TickZoom.Api
 			get { return (int) (elapsed * SecondsPerDay); }
 		}
 		
-		public int TotalMilliseconds {
-			get { return (int) (elapsed * MillisecondsPerDay); }
+		public long TotalMilliseconds {
+			get { return (long) (elapsed * MillisecondsPerDay); }
 		}
 		
 		public int TotalHours {
